Backend/requirements.txt
setuptools
uvicorn[standard]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
authlib
itsdangerous
motor==3.6.0
pydantic[email]
asyncio
httpx
fastapi[standard]
websockets
slowapi
redis==4.5.5
pandas
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
fastapi-sso
asgiref
pymongo
fastapi-cache2[redis]
jinja2
pyjwt
nltk==3.8.1
octoai==1.0.2
opencv_contrib_python==4.8.0.74
pdf2image==1.16.3
Pillow==10.3.0
PyPDF2==3.0.1
pytesseract==0.3.10
python-dotenv==1.0.1
rake_nltk==1.0.6
Requests==2.31.0
streamlit==1.34.0
streamlit_lottie==0.0.5
weasyprint==62.1
pangocffi==0.2.0
pydyf==0.10.0
cffi==1.16.0
html5lib==1.1
tinycss==0.4
cssselect==1.2.0
pyphen==0.15.0
fonttools==4.51.0